requirements.txt
apscheduler
flask
numpy==1.16.4
matplotlib==3.0.3
scipy==1.2.1
setuptools==41.0.1
pyorbital==1.5.0
geographiclib==1.49
tifffile==2019.5.22
Pillow==6.0.0
python_dateutil==2.8.0
attrs==17.4.0
cycler==0.10.0
pluggy==0.6.0
py==1.5.2
pykep==2.3
pyparsing==2.2.0
pytest==3.4.0
pytz==2018.3
six==1.11.0

